v3 - reverse inheritance of service now inherits from abstract_resource
